Made all minibosses use ship health for the health bar. small bug fix in GameLevelScene with the scale of the boss health bar

+++ Game Scenes/GameLevelScene.m
@@ -976,7 +976,7 @@ WrapText( const char *text
         
         if (currentWaveType == kWaveType_Boss) {
             
-            [bossHealthBar setScale:Scale2fMake(bossShip.shipHealth / bossShip.shipMaxHealth, 1.0f)];
+            [bossHealthBar setScale:Scale2fMake((float)bossShip.shipHealth / (float)bossShip.shipMaxHealth, 1.0f)];
 
             // Animating the boss onto the screen
             if (!bossShipIsDisplayed && bossShipReadyToAnimate && bossShip) {
